+++ src/hooks/useCursorInfiniteScroll.ts
@@ -2,39 +2,24 @@ import { useCallback, useEffect, useState } from 'react';
 import { useInView } from 'react-intersection-observer';
 
 export interface CursorInfiniteScrollOptions<T> {
-  /** 데이터를 가져오는 함수 */
   fetchData: (cursorId?: number) => Promise<{
     data: T[];
     nextCursorId: number | null;
   }>;
-  /** 자동 로딩 여부 (기본값: true) */
   autoLoad?: boolean;
-  /** 의존성 배열 (이 값들이 변경되면 데이터를 다시 로드) */
   deps?: unknown[];
 }
 
 export interface CursorInfiniteScrollReturn<T> {
-  /** 현재 로드된 모든 데이터 */
   data: T[];
-  /** 로딩 중인지 여부 */
   isLoading: boolean;
-  /** 더 이상 로드할 데이터가 없는지 여부 */
   hasMore: boolean;
-  /** 에러 상태 */
   error: string | null;
-  /** 다음 페이지를 수동으로 로드하는 함수 */
   loadMore: () => Promise<void>;
-  /** 데이터를 새로고침하는 함수 */
   refresh: () => Promise<void>;
-  /** 무한 스크롤을 위한 ref (Intersection Observer에 연결) */
   ref: (node?: Element | null) => void;
 }
 
-/**
- * 커서 기반 무한 스크롤을 위한 커스텀 훅
- * @param options 커서 기반 무한 스크롤 옵션
- * @returns 커서 기반 무한 스크롤 관련 상태와 함수들
- */
 export function useCursorInfiniteScroll<T>({
   fetchData,
   autoLoad = true,
@@ -46,14 +31,18 @@ export function useCursorInfiniteScroll<T>({
   const [error, setError] = useState<string | null>(null);
   const [nextCursorId, setNextCursorId] = useState<number | null>(null);
 
+  // 무한 스크롤 ref (Intersection Observer에 연결)
   const { ref, inView } = useInView({
     threshold: 0.1,
     rootMargin: '100px',
   });
 
+  // 데이터 누적 및 상태 업데이트
   const loadData = useCallback(
     async (cursorId?: number, reset = false) => {
-      if (isLoading) return;
+      if (isLoading) {
+        return;
+      }
 
       setIsLoading(true);
       setError(null);
@@ -62,17 +51,17 @@ export function useCursorInfiniteScroll<T>({
         const result = await fetchData(cursorId);
 
         if (reset) {
-          setData(result.data);
+          setData(result.data); // 새로고침: 기존 데이터 교체
         } else {
-          setData((prev) => [...prev, ...result.data]);
+          setData((prev) => [...prev, ...result.data]); // 기존 댓글 밑에 새 댓글 추가
         }
 
         setNextCursorId(result.nextCursorId);
         setHasMore(result.nextCursorId !== null);
-      } catch (err) {
+      } catch (error_) {
         setError(
-          err instanceof Error
-            ? err.message
+          error_ instanceof Error
+            ? error_.message
             : '데이터를 불러오는 중 오류가 발생했습니다.'
         );
       } finally {
@@ -83,8 +72,10 @@ export function useCursorInfiniteScroll<T>({
   );
 
   const loadMore = useCallback(async () => {
-    if (!hasMore || isLoading) return;
-    await loadData(nextCursorId, false);
+    if (!hasMore || isLoading) {
+      return;
+    }
+    await loadData(nextCursorId ?? undefined, false);
   }, [hasMore, isLoading, nextCursorId, loadData]);
 
   const refresh = useCallback(async () => {
@@ -95,21 +86,24 @@ export function useCursorInfiniteScroll<T>({
     await loadData(undefined, true);
   }, [loadData]);
 
-  // 자동 로딩
+  // 🎯 무한스크롤 자동 로딩 트리거
+  // 사용자가 스크롤해서 감지 요소가 화면에 보이면 자동으로 다음 데이터 로드
   useEffect(() => {
     if (autoLoad && inView && hasMore && !isLoading) {
-      loadMore();
+      loadMore(); // 다음 댓글 페이지 로드
     }
   }, [inView, hasMore, isLoading, autoLoad, loadMore]);
 
-  // 의존성 변경 시 데이터 새로고침
+  // 🔄 의존성 배열 변경 시 자동 새로고침
+  // deps 배열의 값이 변경되면 (예: cardId 변경) 자동으로 데이터 새로고침
   useEffect(() => {
     if (deps.length > 0) {
       refresh();
     }
-  }, deps);
+  }, [deps, refresh]);
 
-  // 초기 데이터 로드
+  // 🚀 컴포넌트 마운트 시 초기 데이터 로드
+  // 컴포넌트가 처음 렌더링될 때 자동으로 첫 번째 데이터 로드
   useEffect(() => {
     if (autoLoad) {
       refresh();
