--- src/hooks/useInfiniteScroll.ts
@@ -1,143 +1,0 @@
-import { useCallback, useEffect, useState } from 'react';
-import { useInView } from 'react-intersection-observer';
-
-export interface InfiniteScrollOptions<T> {
-  /** 데이터를 가져오는 함수 */
-  fetchData: (page: number) => Promise<{
-    data: T[];
-    hasMore: boolean;
-    totalCount?: number;
-  }>;
-  /** 초기 페이지 번호 (기본값: 1) */
-  initialPage?: number;
-  /** 페이지당 아이템 수 (기본값: 10) */
-  pageSize?: number;
-  /** 자동 로딩 여부 (기본값: true) */
-  autoLoad?: boolean;
-  /** 의존성 배열 (이 값들이 변경되면 데이터를 다시 로드) */
-  deps?: unknown[];
-}
-
-export interface InfiniteScrollReturn<T> {
-  /** 현재 로드된 모든 데이터 */
-  data: T[];
-  /** 로딩 중인지 여부 */
-  isLoading: boolean;
-  /** 더 이상 로드할 데이터가 없는지 여부 */
-  hasMore: boolean;
-  /** 에러 상태 */
-  error: string | null;
-  /** 다음 페이지를 수동으로 로드하는 함수 */
-  loadMore: () => Promise<void>;
-  /** 데이터를 새로고침하는 함수 */
-  refresh: () => Promise<void>;
-  /** 무한 스크롤을 위한 ref (Intersection Observer에 연결) */
-  ref: (node?: Element | null) => void;
-  /** 현재 페이지 번호 */
-  currentPage: number;
-  /** 총 아이템 수 */
-  totalCount: number | null;
-}
-
-/**
- * 무한 스크롤을 위한 커스텀 훅
- * @param options 무한 스크롤 옵션
- * @returns 무한 스크롤 관련 상태와 함수들
- */
-export function useInfiniteScroll<T>({
-  fetchData,
-  initialPage = 1,
-  pageSize = 10,
-  autoLoad = true,
-  deps = [],
-}: InfiniteScrollOptions<T>): InfiniteScrollReturn<T> {
-  const [data, setData] = useState<T[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [hasMore, setHasMore] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [currentPage, setCurrentPage] = useState(initialPage);
-  const [totalCount, setTotalCount] = useState<number | null>(null);
-
-  const { ref, inView } = useInView({
-    threshold: 0.1,
-    rootMargin: '100px',
-  });
-
-  const loadData = useCallback(
-    async (page: number, reset = false) => {
-      if (isLoading) return;
-
-      setIsLoading(true);
-      setError(null);
-
-      try {
-        const result = await fetchData(page);
-
-        if (reset) {
-          setData(result.data);
-        } else {
-          setData((prev) => [...prev, ...result.data]);
-        }
-
-        setHasMore(result.hasMore);
-        setTotalCount(result.totalCount ?? null);
-        setCurrentPage(page);
-      } catch (err) {
-        setError(
-          err instanceof Error
-            ? err.message
-            : '데이터를 불러오는 중 오류가 발생했습니다.'
-        );
-      } finally {
-        setIsLoading(false);
-      }
-    },
-    [fetchData, isLoading]
-  );
-
-  const loadMore = useCallback(async () => {
-    if (!hasMore || isLoading) return;
-    await loadData(currentPage + 1, false);
-  }, [hasMore, isLoading, currentPage, loadData]);
-
-  const refresh = useCallback(async () => {
-    setData([]);
-    setHasMore(true);
-    setCurrentPage(initialPage);
-    setError(null);
-    await loadData(initialPage, true);
-  }, [initialPage, loadData]);
-
-  // 자동 로딩
-  useEffect(() => {
-    if (autoLoad && inView && hasMore && !isLoading) {
-      loadMore();
-    }
-  }, [inView, hasMore, isLoading, autoLoad, loadMore]);
-
-  // 의존성 변경 시 데이터 새로고침
-  useEffect(() => {
-    if (deps.length > 0) {
-      refresh();
-    }
-  }, deps);
-
-  // 초기 데이터 로드
-  useEffect(() => {
-    if (autoLoad) {
-      refresh();
-    }
-  }, []);
-
-  return {
-    data,
-    isLoading,
-    hasMore,
-    error,
-    loadMore,
-    refresh,
-    ref,
-    currentPage,
-    totalCount,
-  };
-}
